fix click cross on workspace/window change

+++ src/app/components/navbar/workspace/workspace.component.ts
@@ -7,6 +7,7 @@ import { ConfirmModalComponent } from './confirm-modal/confirm-modal.component';
 import { CreateModalComponent } from './create-modal/create-modal.component';
 import { RenameModalComponent } from './rename-modal/rename-modal.component';
 import { SettingsService } from 'settings';
+import { FormControl } from '@angular/forms';
 
 @Component({
   selector: 'app-workspace',
@@ -18,6 +19,7 @@ export class WorkspaceComponent implements OnInit {
   @Input() dropdownPlacement: NzPlacementType;
 
   activeWorkspaceId: WorkspaceId;
+  formControl = new FormControl();
 
   workspaces: Workspace[] = [];
   isMenuVisible: boolean;
@@ -35,8 +37,10 @@ export class WorkspaceComponent implements OnInit {
       this.workspaces = [...workspaces];
 
       const activeWorkspace = workspaces.find(w => w.isActive);
-      if (activeWorkspace)
+      if (activeWorkspace) {
         this.activeWorkspaceId = activeWorkspace.id;
+        this.formControl.patchValue(activeWorkspace.id);
+      }
     });
   }
 
@@ -71,7 +75,7 @@ export class WorkspaceComponent implements OnInit {
     });
 
     modal.afterClose.subscribe(result => {
-      if (result)
+      if (result && result.confirmed)
         this._workspacesService.deleteWorkspace(id);
     });
   }
@@ -87,6 +91,7 @@ export class WorkspaceComponent implements OnInit {
   switchWorkspace($event) {
     if (this._settingsService.settings.value.autoSave) {
       this.activeWorkspaceId = $event;
+      this.formControl.patchValue($event);
       this._workspacesService.switchWorkspace(this.activeWorkspaceId);
     } else {
       const modal = this._modalService.create({
@@ -99,9 +104,16 @@ export class WorkspaceComponent implements OnInit {
           cancelText: 'No'
         },
       });
-      modal.afterClose.subscribe((needSave) => {
+      modal.afterClose.subscribe((result: any) => {
+        if (!result) {
+          this.formControl.patchValue(this.activeWorkspaceId);
+          return;
+        }
+        if (result) {
           this.activeWorkspaceId = $event;
-          this._workspacesService.switchWorkspace(this.activeWorkspaceId, needSave);
+          this.formControl.patchValue($event);
+          this._workspacesService.switchWorkspace(this.activeWorkspaceId, result.confirmed);
+        }
       });
     }
   }
@@ -115,7 +127,7 @@ export class WorkspaceComponent implements OnInit {
       nzComponentParams: {
         name: 'Name new workspace',
         blankOption: 'Blank workspace',
-        options: [...this.workspaces.map(item => ({value: item.id, label: item.name}))],
+        options: this.workspaces.map(item => ({ value: item.id, label: item.name })),
       },
     });
